Move remaining web collections code to AC
