+++ apps/web/src/app/admin-console/organizations/collections/collection-badge/collection-name.badge.component.ts
@@ -4,9 +4,14 @@ import { Component, Input } from "@angular/core";
 
 import { CollectionView } from "@bitwarden/admin-console/common";
 
+import { SharedModule } from "../../../../shared/shared.module";
+import { GetCollectionNameFromIdPipe } from "../pipes/get-collection-name.pipe";
+
 @Component({
   selector: "app-collection-badge",
   templateUrl: "collection-name-badge.component.html",
+  standalone: true,
+  imports: [SharedModule, GetCollectionNameFromIdPipe],
 })
 export class CollectionNameBadgeComponent {
   @Input() collectionIds: string[];
