+++ apps/web/src/app/admin-console/organizations/collections/collection-badge/index.ts
@@ -1,0 +1,1 @@
+export * from "./collection-name.badge.component";
