+++ apps/web/src/app/admin-console/organizations/collections/index.ts
@@ -1,0 +1,2 @@
+export * from "./utils";
+export * from "./collection-badge";
