+++ apps/web/src/app/admin-console/organizations/collections/utils/index.ts
@@ -1,0 +1,1 @@
+export * from "./collection-utils";
