+++ apps/web/src/app/vault/individual-vault/pipes/pipes.module.ts
@@ -1,11 +1,10 @@
 import { NgModule } from "@angular/core";
 
-import { GetCollectionNameFromIdPipe } from "./get-collection-name.pipe";
 import { GetGroupNameFromIdPipe } from "./get-group-name.pipe";
 import { GetOrgNameFromIdPipe } from "./get-organization-name.pipe";
 
 @NgModule({
-  declarations: [GetOrgNameFromIdPipe, GetCollectionNameFromIdPipe, GetGroupNameFromIdPipe],
-  exports: [GetOrgNameFromIdPipe, GetCollectionNameFromIdPipe, GetGroupNameFromIdPipe],
+  declarations: [GetOrgNameFromIdPipe, GetGroupNameFromIdPipe],
+  exports: [GetOrgNameFromIdPipe, GetGroupNameFromIdPipe],
 })
 export class PipesModule {}
