+++ apps/web/src/app/vault/individual-vault/vault.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
 
-import { CollectionBadgeModule } from "../../admin-console/organizations/collections/collection-badge/collection-badge.module";
+import { CollectionNameBadgeComponent } from "../../admin-console/organizations/collections";
 import { GroupBadgeModule } from "../../admin-console/organizations/collections/group-badge/group-badge.module";
 import { CollectionDialogComponent } from "../../admin-console/organizations/shared/components/collection-dialog";
 import { LooseComponentsModule, SharedModule } from "../../shared";
@@ -17,7 +17,7 @@ import { ViewComponent } from "./view.component";
     VaultRoutingModule,
     OrganizationBadgeModule,
     GroupBadgeModule,
-    CollectionBadgeModule,
+    CollectionNameBadgeComponent,
     PipesModule,
     SharedModule,
     LooseComponentsModule,
